requirements.txt
python-dotenv
pandas
numpy
pymango
pymongo[srv]
pymongo
certifi
scikit-learn
pyaml
##git -e .